+++ pages/tandemfliegen/index.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <hero-two :hero="hero" />
+    <PageHeader :hero="hero" />
     <product-list :pages="pages" />
     <div class="flex flex-wrap max-w-90 mx-auto py-12 lg:py-24 nuxt-content">
       <h2>Bildergalerie</h2>
